+++ inlang/source-code/comlink-node/worker/not-node.js
@@ -1,1 +1,2 @@
-export const endpoint = undefined
+// eslint-disable-next-line no-undef
+export const endpoint = globalThis