+++ inlang/source-code/paraglide/paraglide-sveltekit/example/src/routes/matchall/page/+page.svelte
@@ -1,4 +1,9 @@
+<script lang="ts">
+
+	import { base } from "$app/paths"
+
+</script>
 <h1>Page</h1>
 
-<a href="/matchall/page/subpage">Subpage</a>
-<a href="/">Home</a>
+<a href="{base}/matchall/page/subpage">Subpage</a>
+<a href="{base}/">Home</a>