+++ inlang/source-code/paraglide/paraglide-sveltekit/example/src/routes/users/[id]/edit/+page.svelte
@@ -1,5 +1,6 @@
 <script>
 	import { enhance } from "$app/forms"
+	import { base } from "$app/paths"
 import { page } from "$app/stores"
     import * as m from "$lib/paraglide/messages.js"
 
@@ -7,7 +8,7 @@ import { page } from "$app/stores"
 </script>
 
 <h1>{m.edit_user({ userId })}</h1>
-<a href="/users/{userId}">{m.users()}</a>
+<a href="{base}/users/{userId}">{m.users()}</a>
 
 <form method="POST" action="?/create" use:enhance>
     <label for="name">Name</label>