+++ inlang/source-code/sdk/slotfile-webapp/src/storage/db-messagebundle.ts
@@ -1,6 +1,5 @@
 import { addRxPlugin } from "rxdb"
 import { RxDBQueryBuilderPlugin } from "rxdb/plugins/query-builder"
-import { createNodeishMemoryFs } from "@lix-js/client"
 import { loadProject } from "@inlang/sdk/v2"
 import { createNewProject } from "@inlang/sdk"
 import http from "isomorphic-git/http/web"
@@ -8,7 +7,7 @@ import http from "isomorphic-git/http/web"
 // NOTE: I use isomorphic git because i went crazy with cors :-/ was faster to spin up a iso proxy
 import git, { pull, add, commit, push, statusMatrix } from "isomorphic-git"
 import { defaultProjectSettings } from "../../../dist/v2/defaultProjectSettings.js"
-
+import type { NodeishFilesystem } from "@lix-js/fs"
 
 addRxPlugin(RxDBQueryBuilderPlugin)
 
@@ -35,12 +34,11 @@ const createAwaitable = () => {
 }
 
 export const openProject = async (
+	fs: NodeishFilesystem,
 	gittoken: string,
 	githubRepo: string,
 	projectPath: string
 ) => {
-	const fs = createNodeishMemoryFs()
-
 	await git.clone({
 		fs: fs,
 		http,