+++ inlang/source-code/sdk/src/v2/rpc-fs/tests/workers/write.ts
@@ -1,0 +1,5 @@
+import { endpoint } from "comlink-node/worker"
+import { getFs } from "../../client.js"
+
+const fs = getFs(endpoint)
+await fs.writeFile("/test.txt", "works!")