+++ lix/packages/server/src/git-proxy.ts
@@ -61,7 +61,6 @@ export async function proxy(request: Request, response: Response, next: NextFunc
 		}
 
 		request.url = "/" + targetUrl
-
 		response.set("Access-Control-Allow-Credentials", "true")
 		response.set("Access-Control-Allow-Headers", "user-agent")
 		response.set("Access-Control-Max-Age", "86400")